requirements.txt
python-decouple
selenium
pandas
sqlalchemy
PyMySQL